Re-fetch SessionResource when session state changes (#6)
Until now the sessionState property of the response was more or less ignored. This adds a listener to HttpJmapApiClient that is used by JmapClient to let SessionClient know about the latest sessionState value. SessionClient will then check if the sessionState value has changed and re-fetch the session resource the next time someone calls SessionClient.get().