save empty query to cache
An earlier commit made the decision not to store empty queries to the cache.
This was essentially a misguided attempt to not trigger endless reloads when
using room and the paging library on Android. (Caused by that implementation
to always write the query state string; thus triggering a reload.)

However empty writes are needed when doing an inital query again after a
canNotCalculateChanges exception.