introduced CorruptCacheException
while CacheConflictExceptions can usually be recovered from and/or only occur
when multiple requests terminate out of order (or Objects had more changes
after piggy bagging on a query request) the new CorruptCacheException only
gets thrown when the Cache is seriously broken and needs to be invalidadet.

Mua will handle that invalidation in the future.