introduce utility method for easy access to required properties
A lot of properties in the Account and Capability are required to be set by the
server. For easier handling it would be nice to be able to access those by their
primitives to avoid null checks. However there is still a slim chance that a
server might forget to send one of those. This commit introduces wrapper
methods around required properties that throw an IllegalStateException when
they are missing.