revert: moveToTrash should not remove label 'Important'
The idea behind the 'moveToTrash should not remove label 'Imporant'' commit
was that we hide emails that are in the 'Trash' mailbox while querying
the 'Important' mailbox; The definition of what is a deleted email would have
changed from 'email that is only in Trash' to 'email that is also in Trash'.

In theory that concept could have been extended to other labels (mailboxes
w/o role); where we always hide emails that are also in Trash.
The benefit would have been that Trashed emails can keep their labels in case
we want to restore them and/or search through trashed emails.

However that breaks JMAP in many ways; 1) we wouldn’t be able to use the
inMailboxesOtherThan filter 2) the mailbox count would have been messed up.