add namespace to using for implicit calls
when `EmailSubmission/set` makes an implicit call to `Email/set` Cyrus requires
us to also add the namespace of `Email/set` to our request.

This may or may not be according to the RFC (see
https://github.com/cyrusimap/cyrus-imapd/issues/2958)

We are chosing the route of annotating the specfic fields that require the
namespace instead of the entire method. That gives use the possibility to in
the future only add the namespace when that field is not null and not empty.

However that approach might be nicer on the wire because it won’t add the
namespace if not strictly needed but it is also harder to cache - we would
have to use reflection on every request.