introduce MockMailServer
MockMailServer generates in memory email threads to be queried for testing purposes.

This commit also refactores some entities to include the id in the builder. This is done
by annotating the constructor instead of the class.

We now also use Filter<T> instead of Filter in Queries. This is necessary to make deserialization work.